phase(S0): move detector core to delivered video_grouper/inference
Behavior-preserving refactor. The game-phase signals + fusion move out of the training
script into the delivered package; training/data_prep/phase_detect.py becomes a thin CLI.

New video_grouper/inference/phase_detector.py (next to field_detector.py) holds the moved
core verbatim: persons/_count_in/player_curve/smooth/segment, whistle_blasts/snap,
ball_restarts, sess + mmss, plus the PLAY_THR/LOUD_GATE/M module globals. Three new wrappers:
  - compute_signals(): player_curve + whistle_blasts + ball_restarts -> cache-shaped dict.
  - fuse_phases(signals): multi_loud + segment + the entire fusion + sanity gate, in TRIMMED
    time (no voff). Returns None for the no-play plateau (was `continue`); kickoff un-clamped.
  - detect_phases(): polygon scaling + compute_signals + fuse_phases for a single video.
No sys.argv, no module-level execution, no prints except under a debug arg. Every EXP-PHASE
comment preserved.

The training CLI keeps the registry loop, all flags, the JSON signal cache (read/write,
--recompute-whistle re-runs only whistle), the .fit.json dump (applies voff) and the
game_state write -- all byte-identical. It imports the core via a package/flat-box shim.

Threading notes (no logic changed): fuse_phases needs the upright source-px field polygon for
its center-circle reference (cxp/cyp), which the cache never stored -- the CLI injects
signals["poly"] each run (it recomputes poly from game.json anyway). compute_signals takes
keyword-only base/gj so the CLI reproduces ball_restarts(d, gj, poly) exactly (dahua per-
segment path included); fuse_phases also returns sig/ko_from_ball/sh_from_ball for the CLI's
_phase_meta. Six percent-format strings and two unused-loop-var unpacks were modernized to
satisfy ruff (video_grouper/ is held to a stricter lint than training/) -- output identical.

Behavior gate: tests/test_phase_fusion.py loads 3 real cached signal sets + their golden
fit.json times from the box and asserts fuse_phases reproduces every boundary (deltas 0.0s).
